+++ src/util/typescriptHelper.ts
@@ -1,16 +1,22 @@
 import { readFileSync } from "fs";
-import { dirname, join } from "path";
+import { dirname, relative } from "path";
 import {
 	CompilerOptions,
 	createProgram,
+	createSemanticDiagnosticsBuilderProgram,
+	createWatchCompilerHost,
+	createWatchProgram,
 	flattenDiagnosticMessageText,
+	formatDiagnostic,
 	getLineAndCharacterOfPosition,
 	getPreEmitDiagnostics,
 	parseConfigFileTextToJson,
 	parseJsonConfigFileContent,
+	SemanticDiagnosticsBuilderProgram,
 	sys,
 	transpileModule,
 	TranspileOutput,
+	WatchOfConfigFile,
 } from "typescript";
 import { OutputChannel, window, workspace } from "vscode";
 import { getSingleFileCompilationCompilerOptions } from "./configurations";
@@ -31,10 +37,7 @@ export function getJsFileName(file: string) {
 export function compileProject(fileNames: string[], options: CompilerOptions): void {
 	// @ts-ignore
 	const _this = compileProject as { outputChannel: OutputChannel };
-
-	if (typeof _this.outputChannel === "undefined") {
-		_this.outputChannel = window.createOutputChannel(`TSC: ${workspace.name}`);
-	}
+	if (typeof _this.outputChannel === "undefined") _this.outputChannel = window.createOutputChannel(`TSC: ${workspace.name}`);
 
 	_this.outputChannel.clear();
 
@@ -64,4 +67,46 @@ export function readConfigFile(configFileName: string): CompilerOptions {
 	const configFileText = readFileSync(configFileName).toString();
 	const result = parseConfigFileTextToJson(configFileName, configFileText);
 	return parseJsonConfigFileContent(result.config, sys, dirname(configFileName)).options;
+}
+
+let watches: { [key: string]: { output: OutputChannel; program: WatchOfConfigFile<SemanticDiagnosticsBuilderProgram> } } = {};
+
+export function watchProject(configPath: string) {
+	if (watches.hasOwnProperty(configPath)) return;
+
+	const host = createWatchCompilerHost(
+		configPath,
+		{},
+		sys,
+		createSemanticDiagnosticsBuilderProgram,
+		(diagnostic) => {
+			let { line, character } = getLineAndCharacterOfPosition(diagnostic.file!, diagnostic.start!);
+			let message = flattenDiagnosticMessageText(diagnostic.messageText, "\n");
+			watches[configPath].output.appendLine(`Error ${diagnostic.file!.fileName} (${line + 1},${character + 1}): ${message}`);
+		},
+		(diagnostic) => {
+			watches[configPath].output.appendLine(
+				formatDiagnostic(diagnostic, {
+					getCanonicalFileName: (path) => path,
+					getCurrentDirectory: sys.getCurrentDirectory,
+					getNewLine: () => sys.newLine,
+				})
+			);
+		}
+	);
+
+	watches[configPath] = {} as any;
+	watches[configPath].output = window.createOutputChannel(`Watch: ${relative(workspace.workspaceFolders![0].uri.fsPath, configPath)}`);
+	watches[configPath].program = createWatchProgram(host);
+}
+
+export function getActiveWatches() {
+	return Object.keys(watches);
+}
+
+export function destroyWatch(configPath: string) {
+	if (!watches.hasOwnProperty(configPath)) return;
+	watches[configPath].program.close();
+	watches[configPath].output.dispose();
+	delete watches[configPath];
 }