+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Change Log
 
 ## [1.0.0]
+
+-   single compilation
+-   live view
+-   compile project
+-   watch project